#1182 - add adminToolTable, move form into page
