+++ src/frontend/src/pages/AdminToolsPage/FinanceConfig/VendorsTable.tsx
@@ -1,4 +1,4 @@
-import { TableRow, TableCell, Paper, Table, TableBody, TableContainer, TableHead, Typography, Box } from '@mui/material';
+import { TableRow, TableCell, Typography, Box } from '@mui/material';
 import LoadingIndicator from '../../../components/LoadingIndicator';
 import { useGetAllVendors } from '../../../hooks/finance.hooks';
 import { datePipe } from '../../../utils/pipes';
@@ -6,6 +6,7 @@ import ErrorPage from '../../ErrorPage';
 import { NERButton } from '../../../components/NERButton';
 import { useState } from 'react';
 import CreateVendorModal from './CreateVendorModal';
+import AdminToolTable from '../AdminToolTable';
 
 const VendorsTable = () => {
   const { data: vendors, isLoading: vendorIsLoading, isError: vendorIsError, error: vendorError } = useGetAllVendors();
@@ -20,7 +21,7 @@ const VendorsTable = () => {
 
   const vendorTableRows = vendors.map((vendor) => (
     <TableRow>
-      <TableCell align="center" sx={{ border: '2px solid black' }}>
+      <TableCell align="left" sx={{ border: '2px solid black' }}>
         {datePipe(vendor.dateCreated)}
       </TableCell>
       <TableCell sx={{ border: '2px solid black' }}>{vendor.name}</TableCell>
@@ -31,30 +32,7 @@ const VendorsTable = () => {
     <Box>
       <CreateVendorModal showModal={createModalShow} handleClose={() => setCreateModalShow(false)} />
       <Typography variant="subtitle1">Registered Vendors</Typography>
-      <TableContainer component={Paper}>
-        <Table sx={{ border: '2px solid black' }}>
-          <TableHead>
-            <TableRow>
-              <TableCell
-                align="center"
-                sx={{
-                  fontSize: '16px',
-                  fontWeight: 600,
-                  width: '25%',
-                  border: '2px solid black'
-                }}
-                itemType="date"
-              >
-                Date Registered
-              </TableCell>
-              <TableCell align="left" sx={{ fontSize: '16px', fontWeight: 600 }}>
-                Vendor Name
-              </TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>{vendorTableRows}</TableBody>
-        </Table>
-      </TableContainer>
+      <AdminToolTable columns={[{ name: 'Date Registered' }, { name: 'Vendor Name' }]} rows={vendorTableRows} />
       <Box sx={{ display: 'flex', justifyContent: 'right', marginTop: '10px' }}>
         <NERButton
           variant="contained"
